+++ src/OF_HARB.cpp
@@ -245,6 +245,8 @@ void FirmHarbor::next_day()
 	else
 		process_queue();
 
+	Ambition::Building::processProductionQueue(this);
+
 	//-*********** simulate ship movement ************-//
 	//if(build_unit_id==0)
 	//	build_ship(UNIT_CARAVEL, 0);
@@ -825,6 +827,8 @@ static void i_disp_queue_button(ButtonCustom *button, int repaintBody)
 	if( harbor->build_unit_id == unitId)
 		queuedCount++;
 
+	queuedCount = Ambition::Building::enqueuedProductionCount(harbor, queuedCount, unitId);
+
 	font_mid.center_put( x1+3, y1+3, x2-3, y2-3, misc.format(queuedCount), 1);
 }
 //--------- End of static function i_disp_queue_button ---------//
@@ -876,6 +880,16 @@ int FirmHarbor::detect_build_menu()
 
 			if( rc==1 )		// left button
 			{
+				if (Ambition::Building::enqueueProduction(this, unitId, createRemoveAmount)) {
+					se_ctrl.immediate_sound("TURN_ON");
+					if (quitFlag) {
+						info.disp();
+					} else {
+						info.update();
+					}
+					return 1;
+				}
+
 				if( remote.is_enable() )
 				{
 					// packet structure : <firm recno> <unit Id> <amount>
@@ -892,6 +906,16 @@ int FirmHarbor::detect_build_menu()
 			}
 			else 				// right button - remove queue
 			{
+				if (Ambition::Building::dequeueProduction(this, unitId, createRemoveAmount)) {
+					se_ctrl.immediate_sound("TURN_OFF");
+					if (quitFlag) {
+						info.disp();
+					} else {
+						info.update();
+					}
+					return 1;
+				}
+
 				if( remote.is_enable() )
 				{
 					// packet structure : <firm recno> <unit Id> <amount>
@@ -956,6 +980,12 @@ int FirmHarbor::detect_build_menu()
 
 	if( queue_ship_selected>=0 && ISKEY(KEYEVENT_MANUF_QUEUE_ADD) )
 	{
+		if (Ambition::Building::enqueueProduction(this, button_unit_id[queue_ship_selected], 1)) {
+			se_ctrl.immediate_sound("TURN_ON");
+			info.update();
+			return 1;
+		}
+
 		if( remote.is_enable() )
 		{
 			// packet structure : <firm recno> <unit Id>
@@ -974,6 +1004,12 @@ int FirmHarbor::detect_build_menu()
 
 	if( queue_ship_selected>=0 && ISKEY(KEYEVENT_MANUF_QUEUE_ADD_BATCH) )
 	{
+		if (Ambition::Building::enqueueProduction(this, button_unit_id[queue_ship_selected], HARBOR_BUILD_BATCH_COUNT)) {
+			se_ctrl.immediate_sound("TURN_ON");
+			info.update();
+			return 1;
+		}
+
 		if( remote.is_enable() )
 		{
 			// packet structure : <firm recno> <unit Id>
@@ -992,6 +1028,12 @@ int FirmHarbor::detect_build_menu()
 
 	if( queue_ship_selected>=0 && ISKEY(KEYEVENT_MANUF_QUEUE_REMOVE) )
 	{
+		if (Ambition::Building::dequeueProduction(this, button_unit_id[queue_ship_selected], 1)) {
+			se_ctrl.immediate_sound("TURN_ON");
+			info.update();
+			return 1;
+		}
+
 		if( remote.is_enable() )
 		{
 			// packet structure : <firm recno> <unit Id>
@@ -1010,6 +1052,12 @@ int FirmHarbor::detect_build_menu()
 
 	if( queue_ship_selected>=0 && ISKEY(KEYEVENT_MANUF_QUEUE_REMOVE_BATCH) )
 	{
+		if (Ambition::Building::dequeueProduction(this, button_unit_id[queue_ship_selected], HARBOR_BUILD_BATCH_COUNT)) {
+			se_ctrl.immediate_sound("TURN_ON");
+			info.update();
+			return 1;
+		}
+
 		if( remote.is_enable() )
 		{
 			// packet structure : <firm recno> <unit Id>